+++ app/interview/VapiInterviewBotClient.tsx
@@ -8,16 +8,12 @@ import { saveInterviewCallId } from "@/app/services/firestoreUser";
 import { auth, db } from "@/app/firebase/firebaseConfig";
 import { doc, getDoc } from "firebase/firestore";
 import { useSearchParams } from 'next/navigation';
-import { useRouter } from 'next/navigation';
 
-import { FiCheckCircle, FiAlertCircle, FiInfo, FiVolumeX, FiVolume2 } from "react-icons/fi";
+import { FiPhoneCall, FiPhoneOff, FiMic, FiMicOff, FiCheckCircle, FiAlertCircle, FiInfo, FiCamera, FiVolumeX, FiVolume2 } from "react-icons/fi";
 
-// NEW: Import the extracted component
-import InterviewRecordingAndFaceDetection from "@/components/InterviewRecordingAndFaceDetection";
+const publicKey = process.env.NEXT_PUBLIC_VAPI_PUBLIC_KEY || "";
 
-const publicKey = process.env.NEXT_PUBLIC_VAPI_PUBLIC_KEY || "abdf35fc-b26c-40e9-9244-0a97666c4b48";
-
-export default function VapiInterviewBotClient() {
+export default function VapiInterviewBotClient() { // Renamed the function
   const [isCallActive, setIsCallActive] = useState<boolean>(false);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
@@ -28,29 +24,15 @@ export default function VapiInterviewBotClient() {
   const [statusMessage, setStatusMessage] = useState<string>("");
   const [hasInterviewBeenTaken, setHasInterviewBeenTaken] = useState<boolean>(false);
 
-  // Video recording states and refs - managed here, passed to child
-  const [recording, setRecording] = useState<boolean>(false);
-  const [recordTime, setRecordTime] = useState<number>(0);
-  const [faceCount, setFaceCount] = useState<number>(0); // State for face count, passed to child
-  const [uploading, setUploading] = useState<boolean>(false);
-  const [userEmail, setUserEmail] = useState<string | null>(null);
-
   const vapiRef = useRef<Vapi | null>(null);
   const isCallActiveRef = useRef<boolean>(isCallActive);
   const currentCallIdRef = useRef<string | null>(null);
-  const localVideoRef = useRef<HTMLVideoElement>(null); // This ref is still managed here
-  const mediaRecorderRef = useRef<MediaRecorder | null>(null); // Still managed here for callbacks
-  const recordedChunksRef = useRef<Blob[]>([]); // Still managed here for data
-  const timerRef = useRef<NodeJS.Timeout | null>(null); // Still managed here for timer logic
+  const localVideoRef = useRef<HTMLVideoElement>(null);
 
   const [currentJobId, setCurrentJobId] = useState<string | null>(null);
 
+  // useSearchParams is now safely inside a Client Component
   const searchParams = useSearchParams();
-  const router = useRouter();
-
-  // State for pre-checks (not fully implemented in UI but good to have)
-  const [cameraAvailable, setCameraAvailable] = useState<boolean | null>(null);
-  const [microphoneAvailable, setMicrophoneAvailable] = useState<boolean | null>(null);
 
   useEffect(() => {
     const jobIdFromParams = searchParams.get('jobId');
@@ -62,25 +44,12 @@ export default function VapiInterviewBotClient() {
     }
   }, [searchParams]);
 
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged(user => {
-      if (user) {
-        setUserEmail(user.email);
-      } else {
-        setUserEmail(null);
-      }
-    });
-    return () => unsubscribe();
-  }, []);
-
   const handleCallEnd = useCallback(async () => {
     setIsCallActive(false);
     setIsLoading(false);
     setErrorMessage(null);
     setIsBotSpeaking(false);
     setStatusMessage("Interview ended.");
-
-    stopRecording(); // Call the local stopRecording function
 
     if (localVideoRef.current && localVideoRef.current.srcObject) {
       (localVideoRef.current.srcObject as MediaStream).getTracks().forEach(track => track.stop());
@@ -141,7 +110,6 @@ export default function VapiInterviewBotClient() {
       console.error("[VAPI Event] Error:", e);
       currentCallIdRef.current = null;
       setIsBotSpeaking(false);
-      stopRecording(); // Ensure recording stops on error
     };
 
     const handleSpeechStart = () => {
@@ -178,7 +146,6 @@ export default function VapiInterviewBotClient() {
         }
         currentCallIdRef.current = null;
       }
-      stopRecording(); // Ensure recording stops on unmount
     };
   }, [publicKey, handleCallEnd]);
 
@@ -247,94 +214,6 @@ export default function VapiInterviewBotClient() {
       }
     }
   }, [isMuted, isCallActive, isBotSpeaking]);
-
-  // --- Video Recording Logic (now methods, not effects) ---
-
-  // Start recording (called by child component)
-  const startRecording = (stream: MediaStream) => {
-    recordedChunksRef.current = [];
-    try {
-      mediaRecorderRef.current = new MediaRecorder(stream, { mimeType: 'video/webm; codecs=vp9' });
-    } catch (error) {
-      console.error("Error creating MediaRecorder:", error);
-      setErrorMessage("Failed to create video recorder. Your browser might not support required codecs.");
-      return;
-    }
-
-    mediaRecorderRef.current.ondataavailable = (event) => {
-      if (event.data.size > 0) {
-        recordedChunksRef.current.push(event.data);
-      }
-    };
-
-    mediaRecorderRef.current.onstop = () => {
-      const blob = new Blob(recordedChunksRef.current, { type: 'video/webm' });
-      console.log("Recording stopped. Blob created:", blob);
-      if (blob.size > 0) {
-        uploadVideo(blob);
-      } else {
-        console.warn("Recorded blob is empty, not uploading.");
-      }
-      recordedChunksRef.current = [];
-    };
-
-    setRecording(true);
-    setRecordTime(0);
-    mediaRecorderRef.current.start();
-    timerRef.current = setInterval(() => setRecordTime((t) => t + 1), 1000);
-    console.log("Recording started.");
-  };
-
-  // Stop recording (called by child component or handleCallEnd)
-  const stopRecording = useCallback(() => {
-    if (mediaRecorderRef.current && mediaRecorderRef.current.state === "recording") {
-      mediaRecorderRef.current.stop();
-      setRecording(false);
-      if (timerRef.current) clearInterval(timerRef.current);
-      timerRef.current = null;
-      console.log("Recording stopped by function call.");
-    }
-  }, []);
-
-  // Upload blob to backend
-  const uploadVideo = async (blob: Blob) => {
-    if (!userEmail) {
-      console.error("Cannot upload video: User email not available.");
-      return;
-    }
-    setUploading(true);
-    const formData = new FormData();
-    formData.append("video", blob, `interview_${Date.now()}.webm`);
-    formData.append("email", userEmail);
-
-    console.log("Uploading video in background...");
-
-    try {
-      const res = await fetch("http://16.171.16.5/api/interview/upload/", {
-        method: "POST",
-        body: formData,
-      });
-
-      if (res.ok) {
-        console.log("Interview uploaded successfully in background!");
-        router.push("/dashboard");
-      } else {
-        let errorText;
-        try {
-          const errorData = await res.json();
-          errorText = JSON.stringify(errorData);
-        } catch {
-          errorText = await res.text();
-        }
-        console.error("Background upload failed:", errorText);
-      }
-    } catch (error) {
-      console.error("Network error during background upload:", error);
-    } finally {
-      setUploading(false);
-    }
-  };
-
 
   const interviewAssistantConfig = useCallback(
     (): CreateAssistantDTO => ({
@@ -424,8 +303,6 @@ export default function VapiInterviewBotClient() {
         localVideoRef.current.play();
       }
       console.log("[Permissions] Microphone and Camera permission granted.");
-      setCameraAvailable(true);
-      setMicrophoneAvailable(true);
 
       console.log("[DEBUG] Attempting Vapi.start() call...");
       const call = await vapiRef.current.start(interviewAssistantConfig());
@@ -442,11 +319,10 @@ export default function VapiInterviewBotClient() {
         setStatusMessage("Failed to start call.");
         setIsLoading(false);
         if (localVideoRef.current && localVideoRef.current.srcObject) {
-          (localVideoRef.current.srcObject as MediaStream).getTracks().forEach(track => track.stop());
-          localVideoRef.current.srcObject = null;
+            (localVideoRef.current.srcObject as MediaStream).getTracks().forEach(track => track.stop());
+            localVideoRef.current.srcObject = null;
         }
         currentCallIdRef.current = null;
-        stopRecording();
       }
     } catch (error: any) {
       console.error("[Start Call] Failed to initiate Vapi interview:", error);
@@ -456,16 +332,13 @@ export default function VapiInterviewBotClient() {
       );
       setStatusMessage(`Failed to start: ${error.message || "Unknown error"}`);
       currentCallIdRef.current = null;
-      setCameraAvailable(false);
-      setMicrophoneAvailable(false);
-      stopRecording();
     }
   };
 
   const handleStopCall = () => {
     if (!vapiRef.current) {
-      console.warn("[Stop Call] Vapi instance not available to stop.");
-      return;
+        console.warn("[Stop Call] Vapi instance not available to stop.");
+        return;
     }
     setIsLoading(true);
     setStatusMessage("Ending call...");
@@ -475,8 +348,8 @@ export default function VapiInterviewBotClient() {
 
   const handleToggleMute = () => {
     if (!vapiRef.current) {
-      console.warn("[Toggle Mute] Vapi instance not available.");
-      return;
+        console.warn("[Toggle Mute] Vapi instance not available.");
+        return;
     }
     const currentlyMuted = vapiRef.current.isMuted();
     vapiRef.current.setMuted(!currentlyMuted);
@@ -485,55 +358,100 @@ export default function VapiInterviewBotClient() {
   };
 
   return (
-    <div className="flex flex-col md:flex-row items-center justify-center min-h-screen p-6 space-y-6 md:space-y-0 md:space-x-8 bg-gray-50"> {/* Increased space-x for symmetry */}
-      {/* AI Bot Section */}
-      <div className="flex flex-col items-center justify-between p-6 md:p-8 min-h-[450px] w-full max-w-md rounded-2xl bg-white shadow-lg border border-gray-200">
+    <div className="flex flex-col md:flex-row items-center justify-center min-h-screen p-6 space-y-6 md:space-y-0 md:space-x-4" style={{ backgroundColor: '#FAFAFC' }}>
+      <div className="flex flex-col items-center justify-between p-6 md:p-8 min-h-[450px] w-full max-w-md rounded-2xl">
         <div className="flex flex-col items-center">
-          <div className={`w-60 h-60 rounded-full overflow-hidden border-4 border-blue-500 shadow-md flex items-center justify-center mb-4 bg-white transform transition-all duration-500
-                ${isBotSpeaking ? 'ring-4 ring-blue-400 ring-opacity-75 animate-pulse-light scale-[1.03] shadow-blue-500/60' : 'hover:scale-[1.01]'}`}>
-            <img src="/images/bot.jpg" alt="AI Bot" className="object-cover w-full h-full" />
-          </div>
+            <div className={`w-60 h-60 rounded-full overflow-hidden border-4 border-blue-500 shadow-md flex items-center justify-center mb-4 bg-white transform transition-all duration-500
+                               ${isBotSpeaking ? 'ring-4 ring-blue-400 ring-opacity-75 animate-pulse-light scale-[1.03] shadow-blue-500/60' : 'hover:scale-[1.01]'}`}>
+                <img src="/images/bot.jpg" alt="AI Bot" className="object-cover w-full h-full" />
+            </div>
         </div>
 
         <div className="mt-auto w-full flex flex-col items-center space-y-3">
           <p className={`text-sm font-medium p-2 rounded-lg border flex items-center justify-center text-center max-w-xs mx-auto
             ${errorMessage ? 'text-red-600 bg-red-100 border-red-200' :
               isScanningResume ? 'text-yellow-700 bg-yellow-100 border-yellow-200 animate-pulse' :
-                isCallActive ? (isMuted ? 'text-orange-700 bg-orange-100 border-orange-200' : (isBotSpeaking ? 'text-blue-700 bg-blue-100 border-blue-200' : 'text-green-700 bg-green-100 border-green-200')) :
-                  'text-green-700 bg-green-100 border-green-200'
+              isCallActive ? (isMuted ? 'text-orange-700 bg-orange-100 border-orange-200' : (isBotSpeaking ? 'text-blue-700 bg-blue-100 border-blue-200' : 'text-green-700 bg-green-100 border-green-200')) :
+              'text-green-700 bg-green-100 border-green-200'
             }`}>
             {errorMessage ? <FiAlertCircle className="w-4 h-4 mr-2" /> :
-              isScanningResume ? <FiInfo className="w-4 h-4 mr-2" /> :
-                isMuted && isCallActive ? <FiVolumeX className="w-4 h-4 mr-2" /> :
-                  isBotSpeaking && isCallActive ? <FiVolume2 className="w-4 h-4 mr-2" /> :
-                    isCallActive ? <FiCheckCircle className="w-4 h-4 mr-2" /> :
-                      <FiInfo className="w-4 h-4 mr-2" />}
+             isScanningResume ? <FiInfo className="w-4 h-4 mr-2" /> :
+             isMuted && isCallActive ? <FiVolumeX className="w-4 h-4 mr-2" /> :
+             isBotSpeaking && isCallActive ? <FiVolume2 className="w-4 h-4 mr-2" /> :
+             isCallActive ? <FiCheckCircle className="w-4 h-4 mr-2" /> :
+             <FiInfo className="w-4 h-4 mr-2" />}
             {statusMessage}
           </p>
         </div>
       </div>
 
-      {/* Video Recording & Face Detection Component with Integrated Controls */}
-      <InterviewRecordingAndFaceDetection
-        isCallActive={isCallActive}
-        userEmail={userEmail}
-        onRecordingStart={startRecording}
-        onRecordingStop={stopRecording}
-        recording={recording}
-        recordTime={recordTime}
-        faceCount={faceCount}
-        setFaceCount={setFaceCount}
-        localVideoRef={localVideoRef}
-        isLoading={isLoading}
-        isMuted={isMuted}
-        hasInterviewBeenTaken={hasInterviewBeenTaken}
-        isScanningResume={isScanningResume}
-        publicKey={publicKey}
-        currentJobId={currentJobId}
-        onStartCall={handleStartCall}
-        onStopCall={handleStopCall}
-        onToggleMute={handleToggleMute}
-      />
+      <div className="flex flex-col items-center justify-between p-6 md:p-8 min-h-[450px] w-full max-w-md rounded-2xl">
+        <div className="w-full max-w-4xl aspect-[16/12] bg-gray-100 rounded-lg overflow-hidden relative border border-gray-300 shadow-sm flex items-center justify-center">
+          <video ref={localVideoRef} autoPlay playsInline muted className="absolute inset-0 w-full h-full object-cover rounded-lg"></video>
+          {!isCallActive && (
+            <div className="absolute inset-0 flex flex-col items-center justify-center bg-gray-900 bg-opacity-70 text-gray-300 text-base">
+              <FiCamera className="w-16 h-16 mb-2 opacity-70" />
+              <p>Camera Preview</p>
+            </div>
+          )}
+        </div>
+
+        <div className="flex space-x-4 mt-auto w-full justify-center">
+          {!isCallActive ? (
+            <button
+              onClick={handleStartCall}
+              disabled={isLoading || isScanningResume || !publicKey || !currentJobId || hasInterviewBeenTaken}
+              className={`flex items-center justify-center w-16 h-16 rounded-full text-white text-xl transition duration-300 ease-in-out transform hover:scale-105 shadow-lg
+                ${
+                  isLoading || isScanningResume || !publicKey || !currentJobId || hasInterviewBeenTaken
+                    ? "bg-gray-400 cursor-not-allowed"
+                    : "bg-green-600 hover:bg-green-700 active:scale-95"
+                }`}
+              aria-label="Start Interview Call"
+            >
+              {isLoading ? (
+                <svg className="animate-spin h-6 w-6 text-white" viewBox="0 0 24 24">
+                  <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                  <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                </svg>
+              ) : (
+                <FiPhoneCall className="w-6 h-6" />
+              )}
+            </button>
+          ) : (
+            <>
+              <button
+                onClick={handleStopCall}
+                disabled={isLoading}
+                className={`flex items-center justify-center w-16 h-16 rounded-full text-white text-xl transition duration-300 ease-in-out transform hover:scale-105 shadow-lg
+                  ${isLoading ? "bg-gray-400 cursor-not-allowed" : "bg-red-600 hover:bg-red-700 active:scale-95"}`}
+                aria-label="End Interview Call"
+              >
+                {isLoading ? (
+                  <svg className="animate-spin h-6 w-6 text-white" viewBox="0 0 24 24">
+                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                  </svg>
+                ) : (
+                  <FiPhoneOff className="w-6 h-6" />
+                )}
+              </button>
+              <button
+                onClick={handleToggleMute}
+                className={`flex items-center justify-center w-16 h-16 rounded-full text-white text-xl transition duration-300 ease-in-out transform hover:scale-105 shadow-lg
+                  ${isMuted ? "bg-orange-500 hover:bg-orange-600" : "bg-blue-600 hover:bg-blue-700"}`}
+                aria-label={isMuted ? "Unmute Microphone" : "Mute Microphone"}
+              >
+                {isMuted ? (
+                  <FiMicOff className="w-6 h-6" />
+                ) : (
+                  <FiMic className="w-6 h-6" />
+                )}
+              </button>
+            </>
+          )}
+        </div>
+      </div>
 
       {!publicKey && (
         <p className="absolute bottom-4 left-1/2 transform -translate-x-1/2 text-center text-red-600 text-sm font-medium bg-red-100 p-2 rounded-lg border border-red-200 max-w-sm shadow-sm">
