+++ package.json
@@ -22,8 +22,6 @@
     "@radix-ui/react-separator": "^1.1.4",
     "@radix-ui/react-slot": "^1.2.0",
     "@radix-ui/react-toast": "^1.2.7",
-    "@tensorflow-models/blazeface": "^0.1.0",
-    "@tensorflow/tfjs": "^4.22.0",
     "@vapi-ai/web": "^2.3.0",
     "autoprefixer": "^10.4.21",
     "class-variance-authority": "^0.7.1",
